+++ langs/src/rust.rs
@@ -1,7 +1,7 @@
 pub mod additional_statements;
 pub mod attribute;
+pub mod builder;
 pub mod mapper;
 pub mod property_generator;
 pub mod reserved_words;
-pub mod type_define_generator;
 pub mod type_statement_generator;
